# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Xuthority API",
    "version": "1.0.0",
    "description": "Automatically generated API documentation.",
    "contact": {
      "name": "Xuthority",
      "url": "https://xuthority.com",
      "email": "support@xuthority.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://xuthority.com/terms",
    "xuthority": "https://xuthority.com",
    "xuthorityDev": "https://xuthority-dev.com",
    "xuthorityDevApi": "http://localhost:8081/api/v1",
    "xuthorityDevApiDocs": "http://localhost:8081/api-docs",
    "xuthorityDevApiDocsSwagger": "http://localhost:8081/swagger.json",
    "xuthorityDevApiDocsSwaggerUi": "http://localhost:8081/api-docs",
    "xuthorityDevApiDocsSwaggerUiSwagger": "http://localhost:8081/swagger.json"
  },
  "servers": [
    {
      "url": "http://localhost:8081/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/product-reviews": {
      "post": {
        "summary": "Create a new product review",
        "tags": [
          "Product Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "product",
                  "overallRating",
                  "title",
                  "content"
                ],
                "properties": {
                  "product": {
                    "type": "string",
                    "format": "objectId"
                  },
                  "overallRating": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "title": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 200
                  },
                  "content": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 5000
                  },
                  "subRatings": {
                    "type": "object",
                    "properties": {
                      "easeOfUse": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 7
                      },
                      "customerSupport": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 7
                      },
                      "features": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 7
                      },
                      "pricing": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 7
                      },
                      "technicalSupport": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 7
                      }
                    }
                  },
                  "reviewSource": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Review already exists for this product"
          }
        }
      },
      "get": {
        "summary": "Get all product reviews with filtering and pagination",
        "tags": [
          "Product Reviews"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "description": "Items per page"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search in title and content"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "flagged"
              ]
            },
            "description": "Review status"
          },
          {
            "in": "query",
            "name": "overallRating",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "minimum": 1,
                "maximum": 5
              }
            },
            "description": "Filter by rating stars"
          },
          {
            "in": "query",
            "name": "minRating",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 5
            },
            "description": "Minimum rating"
          },
          {
            "in": "query",
            "name": "maxRating",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 5
            },
            "description": "Maximum rating"
          },
          {
            "in": "query",
            "name": "isVerified",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by verification status"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "submittedAt",
                "publishedAt",
                "overallRating",
                "helpfulVotes.count"
              ]
            },
            "description": "Sort field"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "description": "Sort order"
          },
          {
            "in": "query",
            "name": "mention",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Filter by popular mentions"
          },
          {
            "in": "query",
            "name": "keywords",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Filter by keywords"
          }
        ],
        "responses": {
          "200": {
            "description": "Reviews retrieved successfully"
          }
        }
      }
    },
    "/api/v1/product-reviews/product/{productId}": {
      "get": {
        "summary": "Get reviews for a specific product",
        "tags": [
          "Product Reviews"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "objectId"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          },
          {
            "in": "query",
            "name": "overallRating",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "minimum": 1,
                "maximum": 5
              }
            }
          },
          {
            "in": "query",
            "name": "isVerified",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "submittedAt",
                "publishedAt",
                "overallRating",
                "helpfulVotes.count"
              ]
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "in": "query",
            "name": "mention",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Filter by popular mentions"
          },
          {
            "in": "query",
            "name": "keywords",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Filter by keywords"
          }
        ],
        "responses": {
          "200": {
            "description": "Product reviews retrieved successfully"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
    "/api/v1/product-reviews/{id}": {
      "get": {
        "summary": "Get a single product review by ID",
        "tags": [
          "Product Reviews"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "objectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review retrieved successfully"
          },
          "404": {
            "description": "Review not found"
          }
        }
      },
      "put": {
        "summary": "Update a product review (owner only)",
        "tags": [
          "Product Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "objectId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "overallRating": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "title": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 200
                  },
                  "content": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 5000
                  },
                  "subRatings": {
                    "type": "object",
                    "properties": {
                      "easeOfUse": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 7
                      },
                      "customerSupport": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 7
                      },
                      "features": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 7
                      },
                      "pricing": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 7
                      },
                      "technicalSupport": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 7
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Can only update own reviews"
          },
          "404": {
            "description": "Review not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a product review (owner or admin only)",
        "tags": [
          "Product Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "objectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Can only delete own reviews"
          },
          "404": {
            "description": "Review not found"
          }
        }
      }
    },
    "/api/v1/product-reviews/{id}/helpful": {
      "post": {
        "summary": "Vote a review as helpful",
        "tags": [
          "Product Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "objectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vote added successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Review not found"
          },
          "409": {
            "description": "Already voted"
          }
        }
      },
      "delete": {
        "summary": "Remove helpful vote from a review",
        "tags": [
          "Product Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "objectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vote removed successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Review or vote not found"
          }
        }
      }
    },
    "/api/v1/product-reviews/{id}/moderate": {
      "patch": {
        "summary": "Moderate a review (admin only)",
        "tags": [
          "Product Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "objectId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "approved",
                      "rejected",
                      "flagged"
                    ]
                  },
                  "moderationNote": {
                    "type": "string",
                    "maxLength": 500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review moderated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "Review not found"
          }
        }
      }
    },
    "/api/v1/product-reviews/product/{productId}/stats": {
      "get": {
        "summary": "Get review statistics for a product",
        "tags": [
          "Product Reviews"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "objectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review statistics retrieved successfully"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
...
  },
...
}